feat: Refactor browser based project.
Breakdown Gruntfile.js into pieces for maintainbility’s sake.
