+++ app/templates/browser/grunt/configs/bump.js
@@ -1,0 +1,15 @@
+module.exports = function (grunt) {
+    grunt.config.set('bump', {
+        options: {
+            files: ['package.json', 'bower.json'],
+            updateConfigs: [],
+            commit: true,
+            commitMessage: 'Release v%VERSION%',
+            commitFiles: ['-a'],
+            createTag: true,
+            tagName: 'v%VERSION%',
+            tagMessage: 'Version %VERSION%',
+            push: false
+        }
+    });
+};
