+++ app/templates/browser/grunt/configs/clean.js
@@ -1,0 +1,6 @@
+module.exports = function (grunt) {
+    grunt.config.set('clean', {
+        dist: ['<%= paths.tmp %>', '<%= paths.dist %>'],
+        server: '<%= paths.tmp %>'
+    });
+};
