+++ app/templates/browser/grunt/configs/concurrent.js
@@ -1,0 +1,16 @@
+module.exports = function (grunt) {
+    grunt.config.set('concurrent', {
+        server: {
+            tasks: ['clean:server', 'compass:server'],
+            options: {
+                logConcurrentOutput: true
+            }
+        },
+        dist: {
+            tasks: ['copy:dist', 'compass:dist'],
+            options: {
+                logConcurrentOutput: true
+            }
+        }
+    });
+};
