+++ app/templates/browser/grunt/configs/htmlmin.js
@@ -1,0 +1,15 @@
+module.exports = function (grunt) {
+    grunt.config.set('htmlmin', {
+        options: {
+            collapseWhitespace: true
+        },
+        dist: {
+            files: [{
+                expand: true,
+                cwd: '<%= paths.dist %>',
+                src: ['**/*.html'],
+                dest: '<%= paths.dist %>'
+            }]
+        }
+    });
+};
