+++ app/templates/browser/grunt/configs/imagemin.js
@@ -1,0 +1,14 @@
+module.exports = function (grunt) {
+    grunt.config.set('imagemin', {
+        dist: {
+            files: [{
+                expand: true,
+                progressive: true,
+                interlace: true,
+                cwd: '<%= paths.dist %>/images',
+                src: '**/*.{png,jpg,jpeg}',
+                dest: '<%= paths.dist %>/images'
+            }]
+        }
+    });
+};
