+++ app/templates/browser/grunt/configs/jshint.js
@@ -1,0 +1,8 @@
+module.exports = function (grunt) {
+    grunt.config.set('jshint', {
+        options: {
+            jshintrc: '.jshintrc'
+        },
+        test: ['<%= paths.app %>/javascripts/**/*.js']
+    });
+};
