+++ app/templates/browser/grunt/configs/rev.js
@@ -1,0 +1,13 @@
+module.exports = function (grunt) {
+    grunt.config.set('rev', {
+        dist: {
+            files: {
+                src: [
+                    '<%= paths.dist %>/javascripts/**/*.js',
+                    '<%= paths.dist %>/stylesheets/**/*.css',
+                    '<%= paths.dist %>/images/**/*.{webp,gif,png,jpg,jpeg,ttf,otf}'
+                ]
+            }
+        }
+    });
+};
