+++ app/templates/browser/grunt/configs/uglify.js
@@ -1,0 +1,10 @@
+module.exports = function (grunt) {
+    grunt.config.set('uglify', {
+        options: {
+            compress: {
+                drop_console: true,
+                drop_debugger: true
+            }
+        }
+    });
+};
