+++ app/templates/browser/grunt/configs/usemin.js
@@ -1,0 +1,17 @@
+module.exports = function (grunt) {
+    grunt.config.set('useminPrepare', {
+        html: ['<%= paths.app %>/**/*.html'],
+        options: {
+            dest: '<%= paths.dist %>'
+        }
+    });
+
+    grunt.config.set('usemin', {
+        html: ['<%= paths.dist %>/**/*.html'],
+        css: ['<%= paths.dist %>/stylesheets/**/*.css'],
+        options: {
+            dirs: ['<%= paths.dist %>'],
+            assetsDirs: ['<%= paths.dist %>']
+        }
+    });
+};
