+++ app/templates/browser/grunt/tasks/serve.js
@@ -1,0 +1,9 @@
+module.exports = function (grunt) {
+    grunt.registerTask('serve', [
+        'concurrent:server',
+        'connect:server',
+        'karma:server',
+        'open',
+        'watch'
+    ]);
+};
