+++ app/templates/browser/grunt/tasks/test.js
@@ -1,0 +1,11 @@
+module.exports = function (grunt) {
+    grunt.registerTask('test', [
+        'jshint:test',
+        'karma:test'
+    ]);
+
+    grunt.registerTask('test:travis', [
+        'jshint:test',
+        'karma:travis'
+    ]);
+};
