+++ app/templates/browser/grunt/tasks/update.js
@@ -1,0 +1,19 @@
+module.exports = function (grunt) {
+    grunt.registerTask(['update'], [
+        'bump-only:patch',
+        'changelog',
+        'bump-commit'
+    ]);
+
+    grunt.registerTask(['update:minor'], [
+        'bump-only:minor',
+        'changelog',
+        'bump-commit'
+    ]);
+
+    grunt.registerTask(['update:major'], [
+        'bump-only:major',
+        'changelog',
+        'bump-commit'
+    ]);
+};
